refactor: make sanitizeInstrumentDescription private to MeterImpl
Address review feedback on #959: sanitizeInstrumentDescription was a
top-level internal extension on SdkErrorHandler, which polluted
autocomplete for every class in the module that has one in scope.
Move it to a private method on MeterImpl (which already owns the
SdkErrorHandler it needs), and fold its code-point helpers into the
same file since they're only used there. Coverage for the edge cases
(at-limit, over-limit, supplementary-plane characters) now lives in
UpDownCounterImplTest, exercised through Meter.createDoubleUpDownCounter
since the sanitizer is no longer independently unit-testable.
